utils: Move signature functions to submodule
