+++ tests/items/distro_test.py
@@ -2,7 +2,8 @@ import os
 
 import pytest
 
-from cobbler import enums, utils
+from cobbler import enums
+from cobbler.utils import signatures
 from cobbler.items.distro import Distro
 from tests.conftest import does_not_raise
 
@@ -38,7 +39,7 @@ def test_equality(cobbler_api):
 def test_make_clone(cobbler_api, create_kernel_initrd, fk_kernel, fk_initrd):
     # Arrange
     folder = create_kernel_initrd(fk_kernel, fk_initrd)
-    utils.load_signatures("/var/lib/cobbler/distro_signatures.json")
+    signatures.load_signatures("/var/lib/cobbler/distro_signatures.json")
     distro = Distro(cobbler_api)
     distro.breed = "suse"
     distro.os_version = "sles15generic"
@@ -174,7 +175,7 @@ def test_boot_loaders(cobbler_api, value, expected_exception, expected_result):
 )
 def test_breed(cobbler_api, value, expected_exception):
     # Arrange
-    utils.load_signatures("/var/lib/cobbler/distro_signatures.json")
+    signatures.load_signatures("/var/lib/cobbler/distro_signatures.json")
     distro = Distro(cobbler_api)
 
     # Act
